Refactor calendar components Composed the components to reduce prop drilling.
