+++ hooks/index.ts
@@ -1,6 +1,9 @@
+'use client'
+
+export * from './useCalendarData'
 export * from './useCalendarForm'
 export * from './useFetchSession'
-export * from './useCalendarData'
+export * from './useMobileCalendarData'
 export * from './useStatus'
 export * from './useSessionStatus'
 export * from './useCalendarIntersectionObserver'
